### 2022-08-20 #4687 @arfon
👋 @nicokurtovic – is this submission associated with a publication in a AAS journal?
