### 2022-08-22 #4687 @nicokurtovic
Hi Arfon,
this submission is not associated with a publication in a AAS journal. The code has been used before in a publication n AAS by Pinilla et al. (2022), but  that paper did not describe the code or its functionalities.
