### 2022-08-25 #4687 @nicokurtovic
Hi @arfon ! This is my first submission to JOSS, should I do something else at this stage of the reviewing process?
