+++ README.md
@@ -8,7 +8,7 @@ SIMIO-continuum is a collection of codes designed to help you compare your radia
 <br/>
 <a href="https://joss.theoj.org/papers/99967c47b9d7e13c7066fffa61e88575"><img src="https://joss.theoj.org/papers/99967c47b9d7e13c7066fffa61e88575/status.svg">
 </a>
-<a href='https://simio-continuum.readthedocs.io/en/main/?badge=main'><img src='https://readthedocs.org/projects/eddy/badge/?version=main' alt='Documentation Status' />
+<a href='https://simio-continuum.readthedocs.io/en/main/?badge=latest'><img src='https://readthedocs.org/projects/eddy/badge/?version=latest' alt='Documentation Status' />
 </a>
 <br/>
 </p>
