### 2022-10-24 ba35337
Corrected npy file load
