### 2022-11-17 #4942 @editorialbot
[REVIEW]: SIMIO-continuum: Connecting simulations to interferometric observations
**Submitting author:** @nicokurtovic (<a href="http://orcid.org/0000-0002-2358-4796">Nicolas Kurtovic</a>)
**Repository:** https://github.com/nicokurtovic/SIMIO
**Branch with paper.md** (empty if default branch): main
**Version:** 1.2
