### 2023-01-25 nicokurtovic/SIMIO#1 @adeleplunkett
Capitalize SIMIO in line 21 of simio_ms2ascii.py
Running the code, I think I found a simple typographical error (but that leads the code to crash, if following the example):
line 21 of simio_ms2ascii.py should be: if current_dir[-6:] != 'SIMIO/': (in current version, simio is in lower case)

This is an issue found when reviewing the code at https://github.com/openjournals/joss-reviews/issues/4942
