### 2023-01-30 nicokurtovic/SIMIO#2 @adeleplunkett
Please add a "statement of need" for the JOSS submission
For the JOSS review (https://github.com/openjournals/joss-reviews/issues/4942), please add a specific "statement of need" section. According to https://joss.readthedocs.io/en/latest/submitting.html, _A Statement of need section that clearly illustrates the research purpose of the software and places it in the context of related work._

I think that you include this implicitly in your documentation, but it can be clarified, which would help the reader for context.
