### 2023-01-30 nicokurtovic/SIMIO#3 @adeleplunkett
"Installation instructions", and especially dependencies, needed
For the JOSS review, https://github.com/openjournals/joss-reviews/issues/4942, please include clear installation instructions.  From the review checklist:

**Installation instructions:** Is there a clearly-stated list of dependencies? Ideally these should be handled with an automated package management solution.

This will help the user greatly.
