### 2023-01-30 nicokurtovic/SIMIO#4 @adeleplunkett
Update CASA reference
The recent reference for CASA should be https://ui.adsabs.harvard.edu/abs/2022PASP..134k4501C/abstract (in addition or in place of the ones you have listed already). See this link: https://casadocs.readthedocs.io/en/latest/notebooks/citing-casa.html

https://github.com/openjournals/joss-reviews/issues/4942
