### 2023-01-30 nicokurtovic/SIMIO#5 @adeleplunkett
Link to tutorial
The link from the Github page to the tutorial should go directly to https://simio-continuum.readthedocs.io/en/main/tutorials/tutorial_1.html, rather than the general documentation.  (This is a minor issue.)

https://github.com/openjournals/joss-reviews/issues/4942
