### 2023-01-30 nicokurtovic/SIMIO#6 @adeleplunkett
Add some explanation of the template file in the Example.
In the “Example” on README.md, it is not clear how to get the exact template file, and where to put it. I guessed, which worked, but another line could help the general user.

https://github.com/openjournals/joss-reviews/issues/4942
