### 2023-01-30 nicokurtovic/SIMIO#7 @adeleplunkett
Please add to "Community Guidelines" for your software
For the JOSS review, please clarify this guideline:

**Community guidelines:** Are there clear guidelines for third parties wishing to 1) Contribute to the software 2) Report issues or problems with the software 3) Seek support

https://github.com/openjournals/joss-reviews/issues/4942
