### 2023-01-30 nicokurtovic/SIMIO#8 @adeleplunkett
In example, directory projects/SolarS_HD163296/images is already filled
In the example project: SolarS_HD163296.tar.gz, the directory projects/SolarS_HD163296/images is already filled.  This might be helpful, but it's not written anywhere in the example about whether the images can be overwritten.  Maybe just a few words for clarity will help.

https://github.com/openjournals/joss-reviews/issues/4942
