### 2023-01-30 nicokurtovic/SIMIO#9 @adeleplunkett
Recommend to update code for compatibility with CASA 6
As correctly noted in the documentation, the code runs with CASA 5 (I tested with CASA 5.6.1-8).  I would strongly recommend to update it for compatibility with CASA 6.

https://casa.nrao.edu/casa_obtaining.shtml

https://github.com/openjournals/joss-reviews/issues/4942
