### 2023-01-30 nicokurtovic/SIMIO#10 @adeleplunkett
Minor edits to paper.md
Minor edits to paper.md:

- Emonts:2020 may need the  @ symbol.
- Convolving with a gaussian —> Gaussian should be capitalized

https://github.com/openjournals/joss-reviews/issues/4942
