### 2023-01-30 nicokurtovic/SIMIO#11 @adeleplunkett
Some information on other tools that do something similar
It is important to put this work in the context of other tools that do something similar.
- [https://mtazzari.github.io/galario/index.html](https://mtazzari.github.io/galario/index.html) seems to be able to generate visibilities of a model image in the same (u,v) coordinates as an observation.  That code is more focused on speeding up the process by allowing use with CPU and GPU.  It is already being used in numerous publications: [https://ui.adsabs.harvard.edu/search/p_=0&q=citations(bibcode%3A2018MNRAS.476.4527T) &sort=date desc%2C bibcode desc](https://ui.adsabs.harvard.edu/search/p_=0&q=citations(bibcode%3A2018MNRAS.476.4527T)%20&sort=date%20desc%2C%20bibcode%20desc). However, I found that SIMIO-continuum makes this process much simpler for the basic case you have presented (i.e. it’s possible to plug in simply a template and image).
- Another case that I am aware of is this one: [https://github.com/onsala-space-observatory/fakeobs](https://github.com/onsala-space-observatory/fakeobs). It does not have extensive user documentation, but it is described as: FAKEOBS is a CASA task to generate model visibilities from already-existing measurement sets.  It is found here: [https://www.oso.nordic-alma.se/software-tools.php](https://www.oso.nordic-alma.se/software-tools.php).
- Something similar can also be done with toolkit calls in CASA using a little care, see: https://casaguides.nrao.edu/index.php?title=Fit_an_arbitrary_sky_(image)_model_to_an_existing_MS

I find your software, and example, to be simple and straight forward to use, so I find value in that.  This list is just to be aware of other tools that may be complementary, and a user may like to know about them.

https://github.com/openjournals/joss-reviews/issues/4942
