### 2023-01-30 #4942 @adeleplunkett
@ivastar, I have completed my review, with issues opened in https://github.com/nicokurtovic/SIMIO.  Overall I find this tool to be useful, the example was straightforward to run, and documentation was sufficient.  I expect this will be a helpful contribution to ALMA users.  I did not test that the output is exactly "correct".

I hope that the author finds the comments constructive.  Please let me know if there are any other contributions needed on my side.

Best regards,
Adele
