### 2023-02-10 nicokurtovic/SIMIO#11 @jeffjennings
Mentioning existing tools that have similar functionality (e.g. galario) should also be in the JOSS paper - to give more context on the state of the field (from the review guidelines - _State of the field: Do the authors describe how this software compares to other commonly-used packages?_) - and to include in the references there (_References: Is the list of references complete..._).

[https://github.com/openjournals/joss-reviews/issues/4942](https://github.com/openjournals/joss-reviews/issues/4942)
