### 2023-02-10 nicokurtovic/SIMIO#2 @jeffjennings
I agree, please make this its own section in the JOSS paper.

[https://github.com/openjournals/joss-reviews/issues/4942](https://github.com/openjournals/joss-reviews/issues/4942)
