### 2023-02-10 nicokurtovic/SIMIO#13 @jeffjennings
Make SIMIO a standalone Python package
I appreciate the code is designed to serve as executable functions inside CASA, and I know this would be a bit of a larger effort, but I think it would be highly beneficial if the code were also available as a standalone Python package (on PyPI and thus installable by pip).

The only major addition to the code's core routines would be a method to interface with measurement sets; you could model this after existing packages (or just make them dependencies and import them) such as [uvplot](https://pypi.org/project/uvplot/) and [visread](https://pypi.org/project/visread/). This would also allow users to use the code with modular CASA (e.g., as is done with [PRIISM](https://github.com/tnakazato/priism#installation-procedure-in-detail)).

[https://github.com/openjournals/joss-reviews/issues/4942](https://github.com/openjournals/joss-reviews/issues/4942)
