### 2023-02-10 nicokurtovic/SIMIO#3 @jeffjennings
I agree; for the installation instructions, you could consider moving steps 1 - 4 in the [docs tutorial 1](https://simio-continuum.readthedocs.io/en/main/tutorials/tutorial_1.html) to the [docs install](https://simio-continuum.readthedocs.io/en/main/content/simio_installation.html).

For the dependencies, a `setup.py` in the top level of the codebase should list all dependencies in the scripts (numpy, etc.).

See #13  for further discussion on both points.

[https://github.com/openjournals/joss-reviews/issues/4942](https://github.com/openjournals/joss-reviews/issues/4942)
