### 2023-02-10 nicokurtovic/SIMIO#7 @jeffjennings
I agree, these would be really useful for code users and for those who want to contribute.

[https://github.com/openjournals/joss-reviews/issues/4942](https://github.com/openjournals/joss-reviews/issues/4942)
