### 2023-02-10 #4942 @jeffjennings
@ivastar, I've (finally) completed my review, thank you for your patience! I agree with each of @adeleplunkett 's points, and I've added a few comments to some of the associated issues in the code repository, as well as opened a couple new issues.

I find the code useful and well-documented. Among the unchecked items on my review checklist, the largest is related to _Installation instructions_ - I would prefer the software to be a standalone python package rather than only a set of routines that are run from within CASA software (I've opened an issue to this end). Once all open issues in the SIMIO repository are addressed, I would be happy to revisit the review checklist. If there's anything else I can provide, please let me know.

Also, in the JOSS paper, if possible could my "@" to @jeffjennings (sorry for misspelling my own tag that's in the current submission)? If useful to include, my orcid ID is 0000-0002-7032-2350.

Thanks again!
