+++ docs/requirements.txt
@@ -4,3 +4,4 @@ jupyter
 nbsphinx
 pygments>=2.4.1
 docutils<0.18
+sphinx-rtd-theme
