+++ docs/requirements.txt
@@ -1,6 +1,7 @@
 sphinx>=1.7.5
 pandoc
 jupyter
+ipython
 nbsphinx
 pygments>=2.4.1
 docutils<0.18
