### 2023-10-15 0ee6105
Update conf.py with extensions
