+++ docs/conf.py
@@ -34,7 +34,9 @@ extensions = [
     'sphinx.ext.coverage',
     'sphinx.ext.napoleon',
     'sphinx.ext.imgmath',
+    'sphinx.ext.todo',
     'sphinx.ext.mathjax',
+    'sphinx.ext.ifconfig'
     'sphinx_rtd_theme',
     'nbsphinx',
 ]
