+++ .readthedocs.yaml
@@ -5,6 +5,9 @@
 # Required
 version: 2
 
+build:
+   os: ubuntu-20.04
+
 # Build documentation in the docs/ directory with Sphinx
 sphinx:
    configuration: docs/conf.py
