+++ .readthedocs.yaml
@@ -7,6 +7,8 @@ version: 2
 
 build:
    os: ubuntu-20.04
+   tools:
+      python: "3.7"
 
 # Build documentation in the docs/ directory with Sphinx
 sphinx:
