+++ .readthedocs.yaml
@@ -5,11 +5,6 @@
 # Required
 version: 2
 
-build:
-   os: ubuntu-20.04
-   tools:
-      python: "3.7"
-
 # Build documentation in the docs/ directory with Sphinx
 sphinx:
    configuration: docs/conf.py
