+++ .readthedocs.yaml
@@ -5,6 +5,11 @@
 # Required
 version: 2
 
+build:
+  os: ubuntu-22.04
+  tools:
+    python: "3.7"
+
 # Build documentation in the docs/ directory with Sphinx
 sphinx:
    configuration: docs/conf.py
@@ -14,7 +19,5 @@ formats:
    - pdf
 
 # Optionally set the version of Python and requirements required to build your docs
-python:
-   version: 3.7
-   install:
+install:
    - requirements: docs/requirements.txt
