+++ .readthedocs.yaml
@@ -19,5 +19,6 @@ formats:
    - pdf
 
 # Optionally set the version of Python and requirements required to build your docs
-install:
-   - requirements: docs/requirements.txt
+python:
+  install:
+  - requirements: docs/requirements.txt
