+++ docs/requirements.txt
@@ -6,3 +6,4 @@ nbsphinx
 pygments>=2.4.1
 docutils<0.18
 sphinx-rtd-theme
+pydata-sphinx-theme
