+++ docs/requirements.txt
@@ -1,9 +1,10 @@
-sphinx>=1.7.5
+sphinx>=3.1
 pandoc
 jupyter
 ipython
 nbsphinx
-pygments>=2.4.1
 docutils<0.18
+sphinx-intl
 sphinx-rtd-theme
 pydata-sphinx-theme
+pygments>=2.4.1
