+++ docs/requirements.txt
@@ -1,5 +1,5 @@
 sphinx>=3.1
-pandoc
+pandoc>=2.1
 jupyter
 ipython
 nbsphinx
