+++ docs/conf.py
@@ -36,7 +36,6 @@ extensions = [
     'sphinx.ext.imgmath',
     'sphinx.ext.todo',
     'sphinx.ext.mathjax',
-    'sphinx.ext.pngmath',
     'sphinx.ext.ifconfig',
     'sphinx_rtd_theme',
     'nbsphinx',
