+++ docs/requirements.txt
@@ -11,3 +11,4 @@ pygments>=2.4.1
 myst_parser
 sphinxext-rediraffe
 sphinx-panels
+sphinx-autobuild
