+++ .readthedocs.yaml
@@ -21,4 +21,4 @@ formats:
 # Optionally set the version of Python and requirements required to build your docs
 python:
   install:
-  - requirements: docs/requirements.txt
+    - requirements: docs/requirements.txt
