+++ docs/conf.py
@@ -10,8 +10,8 @@
 # add these directories to sys.path here. If the directory is relative to the
 # documentation root, use os.path.abspath to make it absolute, like shown here.
 #
-import os
 import sys
+import os
 sys.path.insert(0, os.path.abspath('.'))
 # autodoc_member_order = 'bysource'  # uncomment to not have alphabetical list.
 
