+++ docs/requirements.txt
@@ -2,9 +2,7 @@ sphinx>=3.1
 pandoc>=2.1
 jupyter
 ipython
-ipykernel
-libjs-mathjax
-nbsphinx
+ipykernelnbsphinx
 docutils<0.18
 sphinx-intl
 sphinx-rtd-theme
