+++ docs/conf.py
@@ -38,6 +38,7 @@ extensions = [
     'sphinx.ext.mathjax',
     'sphinx.ext.ifconfig',
     'sphinx_rtd_theme',
+    'sphinx_math_dollar',
     'nbsphinx',
 ]
 
