+++ docs/requirements.txt
@@ -2,7 +2,8 @@ sphinx>=3.1
 pandoc>=2.1
 jupyter
 ipython
-ipykernelnbsphinx
+ipykernel
+nbsphinx
 docutils<0.18
 sphinx-intl
 sphinx-rtd-theme
