### 2023-10-17 #4942 @nicokurtovic
Hi @ivastar. I am actively fixing the comments, I will reply to them after I have implemented all the changes. Currently what is taking me the longest is to fix a problem with readthedocs displaying equations, but I hope I can fix it very soon.
