+++ .readthedocs.yaml
@@ -12,7 +12,7 @@ build:
 
 # Build documentation in the docs/ directory with Sphinx
 sphinx:
-   configuration: docs/conf.py
+   configuration: ./docs/conf.py
 
 # Optionally build your docs in additional formats such as PDF
 formats:
@@ -21,4 +21,4 @@ formats:
 # Optionally set the version of Python and requirements required to build your docs
 python:
   install:
-    - requirements: docs/requirements.txt
+    - requirements: ./docs/requirements.txt
