### 2023-10-17 b34b342
Update conf.py, add mathjax path
