+++ docs/index.rst
@@ -19,6 +19,7 @@ indistinctively.
    content/content_1
    content/content_2
    content/content_3
+   content/content_9
 
 .. toctree::
    :maxdepth: 2
