+++ docs/conf.py
@@ -16,6 +16,7 @@ sys.path.insert(0, os.path.abspath('.'))
 # autodoc_member_order = 'bysource'  # uncomment to not have alphabetical list.
 
 sys.path.append(os.path.abspath('./content/'))
+sys.path.append(os.path.abspath('../../../../../'))
 sys.path.append(os.path.abspath('../../../../'))
 sys.path.append(os.path.abspath('../../../'))
 sys.path.append(os.path.abspath('../../'))
@@ -42,6 +43,7 @@ extensions = [
     'sphinx.ext.todo',
     'sphinx.ext.mathjax',
     'sphinx.ext.ifconfig',
+    'sphinx.ext.viewcode',
     'sphinx_rtd_theme',
     'nbsphinx', 
 ]
