+++ docs/content/content_9.rst
@@ -49,6 +49,7 @@ This is a test. Here is an equation:
 Here is another:
 
 .. math::
+r"""
     :label: This is a label
 
     \nabla^2 f =
@@ -57,6 +58,7 @@ Here is another:
     \frac{1}{r^2 \sin \theta} \frac{\partial f}{\partial \theta}
     \left( \sin \theta \, \frac{\partial f}{\partial \theta} \right) +
     \frac{1}{r^2 \sin^2\theta} \frac{\partial^2 f}{\partial \phi^2}
+"""
 
 You can add a link to equations like the one above :eq:`This is a label` by using ``:eq:``.
 
