### 2023-10-18 868b2b3
Update content_9.rst
