### 2023-10-18 63c00ec
Update .readthedocs.yaml
