+++ .readthedocs.yaml
@@ -12,7 +12,7 @@ build:
 
 # Build documentation in the docs/ directory with Sphinx
 sphinx:
-   builder: singlehtml
+   builder: html
    configuration: ./docs/conf.py
 
 # Optionally build your docs in additional formats such as PDF
