+++ docs/conf.py
@@ -49,11 +49,11 @@ extensions = [
 ]
 
 html_theme = 'sphinx_rtd_theme' 
-imgmath_latex = "pdflatex"
+#imgmath_latex = "pdflatex"
 
 # Mathjax path
-#mathjax_path = 'https://cdn.mathjax.org/mathjax/latest/MathJax.js?config=TeX-AMS-MML_HTMLorMML'
-mathjax_path = 'https://cdn.jsdelivr.net/npm/mathjax@2/MathJax.js?config=TeX-AMS-MML_HTMLorMML'
+mathjax_path = 'https://cdn.mathjax.org/mathjax/latest/MathJax.js?config=TeX-AMS-MML_HTMLorMML'
+#mathjax_path = 'https://cdn.jsdelivr.net/npm/mathjax@2/MathJax.js?config=TeX-AMS-MML_HTMLorMML'
 
 # astropy
 autodoc_mock_imports = ['astropy']
