+++ docs/conf.py
@@ -44,7 +44,7 @@ extensions = [
     'sphinx.ext.mathjax',
     'sphinx.ext.ifconfig',
     'sphinx.ext.viewcode',
-    #'sphinx_rtd_theme',
+    'sphinx_rtd_theme',
     'nbsphinx', 
 ]
 
