+++ .readthedocs.yaml
@@ -8,7 +8,8 @@ version: 2
 build:
   os: ubuntu-22.04
   tools:
-    python: "3.7"
+    #python: "3.7"
+    python: "3.11"
 
 # Build documentation in the docs/ directory with Sphinx
 sphinx:
