+++ docs/requirements.txt
@@ -14,3 +14,8 @@ sphinxext-rediraffe
 sphinx-panels
 sphinx-autobuild
 sphinx-math-dollar
+
+pillow==5.4.1
+mock==1.0.1
+commonmark==0.9.1
+recommonmark==0.5.0
