+++ docs/build/html/index.rst
@@ -1,0 +1,58 @@
+SIMIO-continuum - SIMulatIons to Observations
+===============================
+
+*SIMIO-continuum* is a collection of codes designed to help you compare
+your radiative transfer images with existing ALMA observations. With 
+*SIMIO-continuum* you can generate your synthetic observations with
+different geometries, distances, and even including white noise.
+
+Please check www.nicolaskurtovic.com/simio and the Contents section
+for a detailed description of the package. Although *SIMIO-continuum*
+is the official name, this page uses *SIMIO-continuum* or *SIMIO*
+indistinctively.
+
+.. toctree::
+   :maxdepth: 2
+   :caption: Content
+
+   content/content_0
+   content/content_1
+   content/content_2
+   content/content_3
+   content/content_9
+
+.. toctree::
+   :maxdepth: 2
+   :caption: Information
+   
+   content/api_simio
+   content/simio_installation
+   tutorials/testsimio_0
+   content/considerations
+   content/literature_simio
+
+
+.. toctree::
+   :maxdepth: 2
+   :caption: Tutorials
+   
+   tutorials/tutorial_1
+   tutorials/tutorial_2
+   tutorials/tutorial_3
+   tutorials/tutorial_4
+   tutorials/tutorial_5
+
+
+Support
+-------
+
+If you are having issues, please open a issue on the GitHub page, or send me a
+message to my email address.
+
+
+License
+-------
+
+The project is licensed under the `MIT license`_.
+
+.. _MIT license: https://github.com/nicokurtovic/SIMIO/blob/main/LICENSE
