+++ .readthedocs.yaml
@@ -8,12 +8,12 @@ version: 2
 build:
   os: ubuntu-22.04
   tools:
-    #python: "3.7"
-    python: "3.11"
+    python: "3.7"
+    #python: "3.11"
 
 # Build documentation in the docs/ directory with Sphinx
 sphinx:
-   builder: html
+   builder: simglehtml
    configuration: ./docs/conf.py
 
 # Optionally build your docs in additional formats such as PDF
