+++ .readthedocs.yaml
@@ -13,7 +13,7 @@ build:
 
 # Build documentation in the docs/ directory with Sphinx
 sphinx:
-   builder: singlehtml
+   builder: dirhtml
    configuration: ./docs/conf.py
 
 # Optionally build your docs in additional formats such as PDF
