### 2023-10-31 ad5c9ca
Update .readthedocs.yaml to singlehtml
