+++ .readthedocs.yaml
@@ -13,7 +13,7 @@ build:
 
 # Build documentation in the docs/ directory with Sphinx
 sphinx:
-   builder: html   #singlehtml
+   builder: singlehtml
    configuration: ./docs/conf.py
 
 # Optionally build your docs in additional formats such as PDF
