### 2023-11-02 c427138
Modified geometric handling of visibilities
The function deproject_uv was replaced by handle_uv, which calls sub functions to execute the projection, deprojection, rotation, and shifting of visibilities
