+++ codes/simio_support.py
@@ -23,9 +23,101 @@ from scipy.integrate import trapz, quadrature
 ##############################################################################
 
 
-def deproject_uv(u, v, Re, Im, inc, pa, dRa, dDec, inverse=False):
+def _shift_uv(u, v, Re, Im, dRa, dDec):
     '''
-    Projects and deprojects visibility points based on the input parameters
+    Applies rotation and inclination to the input visibilities.
+    
+    Args:
+        - u: (numpy array) Array containing the ``u`` coordinates of the
+                    visibilities.
+        - v: (numpy array) Array containing the ``v`` coordinates of the
+                    visibilities.
+        - Re: (numpy array) Array containing the ``Real`` part of the
+                    visibilities.
+        - Im: (numpy array) Array containing the ``Imaginary`` part of the
+                    visibilities.
+        - inc: (float) Value to incline the visibilities, in **degrees**.
+        - pa: (float) Value to rotate the visibilities, in **degrees**.
+    '''
+    # Calculate the value of the visibilities
+    intensity = (Re + Im * 1j) * np.exp(2j * np.pi * (u * -dRa + v * -dDec))
+    # Separate in real and imaginary part
+    Re_shift = np.real(intensity)
+    Im_shift = np.imag(intensity)
+    # Return
+    return u, v, Re_shift, Im_shift
+
+
+def _rotate_uv(u, v, pa):
+    '''
+    Applies a rotation matrix to the visibilities, given a rotation angle
+    of 'pa'.
+
+    Args:
+        - u: (numpy array) Array containing the ``u`` coordinates of the
+                    visibilities.
+        - v: (numpy array) Array containing the ``v`` coordinates of the
+                    visibilities.
+        - pa: (float) Value to rotate the visibilities, in **radians**.
+    '''
+    # Calculate rotation components
+    cos_rot = np.cos(pa)
+    sin_rot = np.sin(pa)
+    # Rotate coordinates
+    u_rot = u * cos_rot - v * sin_rot
+    v_rot = u * sin_rot + v * cos_rot
+    # Return
+    return u_rot, v_rot
+
+
+def _deproject_uv(u, v, inc):
+    '''
+    Applies rotation and inclination to the input visibilities.
+    
+    Args:
+        - u: (numpy array) Array containing the ``u`` coordinates of the
+                    visibilities.
+        - v: (numpy array) Array containing the ``v`` coordinates of the
+                    visibilities.
+        - Re: (numpy array) Array containing the ``Real`` part of the
+                    visibilities.
+        - Im: (numpy array) Array containing the ``Imaginary`` part of the
+                    visibilities.
+        - inc: (float) Value to incline the visibilities, in **degrees**.
+        - pa: (float) Value to rotate the visibilities, in **degrees**.
+    '''
+    # Apply inclination
+    u_inc = u * np.cos(inc)
+    # Return
+    return u_inc, v
+
+
+def _project_uv(u, v, inc):
+    '''
+    Applies rotation and inclination to the input visibilities.
+    
+    Args:
+        - u: (numpy array) Array containing the ``u`` coordinates of the
+                    visibilities.
+        - v: (numpy array) Array containing the ``v`` coordinates of the
+                    visibilities.
+        - Re: (numpy array) Array containing the ``Real`` part of the
+                    visibilities.
[…]
