### 2023-11-02 1e80949
Updated change_geom function with handle_uv
