+++ codes/simio_ms2ascii.py
@@ -37,8 +37,6 @@ execfile(current_dir+'codes/simio_clean.py')
 
 def ms_to_ascii(ms_cont, ascii_file, with_flags):
     '''
-    Adapted from Laura Perez
-    
     Takes ms_cont, and writes the visibilities in a file named ``ascii_file``.
     The visibility table will have the structure of (u, v, Real, Imag, weights).
     If you do not want to include flagged data in your visibility table, then
@@ -141,8 +139,6 @@ def ms_to_ascii(ms_cont, ascii_file, with_flags):
 
 def ascii_to_ms(ascii_file, ms_file, new_ms_file):
     '''
-    Adapted from Laura Perez
-    
     Inverse function of ms_to_ascii. It takes a visibility table in a ``.txt``
     or ``.dat`` format, and replaces those visibilities in a ms_file. The
     ``ms_file`` must match exactly the visibility table in number of spw and
@@ -199,164 +195,6 @@ def ascii_to_ms(ascii_file, ms_file, new_ms_file):
     tb.close()
     # Print
     print ('New ms file is: ' + new_ms_file)
-
-
-def _write_temp_uvtables(simobj):
-    '''
-    Write the template uvtable. This is wrapper to write the uvtable of the
-    ``simio_object`` template, to be used when executing *SIMIO* with
-    ``galario``.
-    
-    Args:
-        - simobj: (simio_object) *SIMIO* object containing the synthetic
-                    observation that will be extracted.
-    Returns:
-        Writes the visibility table of each spectral window.
-    '''
-    # Remove previous existent uvtables
-    os.system('rm ' + simobj._source_dir + 'msfiles/'  + simobj._prefix + '_cont_spw*.txt')
-    os.system('rm ' + simobj._source_dir + 'uvtables/' + simobj._prefix + '_cont_spw*.txt')
-    # Write individual uvtables for each spw
-    for i in simobj._spw_temp:
-        # Names
-        ms_file    = simobj._source_dir + 'msfiles/'  + simobj._prefix + '_cont_spw' + str(i) + '.ms'
-        ascii_file = simobj._source_dir + 'uvtables/' + simobj._prefix + '_cont_spw' + str(i) + '.txt'
-        # Delete any previous spw file
-        os.system('rm -rf ' + ms_file)
-        os.system('rm -rf ' + ascii_file)
-        # Split
-        split(vis=simobj._ms_temp, outputvis=ms_file, \
-              keepflags=False, datacolumn='data', spw=i)
-        ms_to_ascii(ms_file, ascii_file, with_flags=True)
-        # Delete temporal msfile
-        os.system('rm -rf ' + ms_file)
-    # Check weight
-    os.system('du -h ' + simobj._source_dir + 'uvtables')
-
-
-def _write_mod_uvtables(simobj):
-    '''
-    Writes the uvtables of the model in ``simio_object``. This function is
-    written to work with the ``galario`` mode of ``simio_object``.
-    
-    Args:
-        - simobj: (simio_object) *SIMIO* object containing the synthetic
-                    observation that will be written.
-    Returns:
-        Writes the visibility table of each spectral window, with the
-        visibilities from the model.
-    '''
-    # Iterate over all spws
-    for i in simobj._spw_temp.astype(str):
-        print (i)
-        # Read uvtable from individual spw
-        ascii_file = simobj._source_dir + 'uvtables/' + simobj._prefix + '_cont_spw'+str(i)+'.txt'
-        uu, vv, RReal, IImag, WWei = np.loadtxt(ascii_file, unpack=True)
-        # C-contiguous
-        uu = np.ascontiguousarray(uu)
-        vv = np.ascontiguousarray(vv)
-        RReal = np.ascontiguousarray(RReal)
[…]
