### 2023-11-02 2721a04
Changed check of current folder
