### 2023-11-02 nicokurtovic/SIMIO#1 @nicokurtovic
Thank you for pointing out this error. The check was initially
