### 2023-11-02 nicokurtovic/SIMIO#1 @nicokurtovic
Thank you for pointing out this error. The check was initially implemented so that SIMIO could be executed from inside a template folder. I have changed it now to a printed warning if the folder is different from 'SIMIO' with  capital letters
