### 2023-11-02 50dc1b8
Removed the test element content 9
Content 9 was included to test the display of equations in the readthedocs webpage. It is not longer needed.
