### 2023-11-02 eaf2910
Deleted the test element docs/content/content_9.rst
