+++ docs/content/content_9.rst
@@ -1,62 +1,0 @@
-Inline Markup
-=============
-
-Paragraphs contain text and may contain inline markup: *emphasis*, **strong emphasis**, ``inline literals``,
-standalone hyperlinks (http://www.python.org), external hyperlinks (Python_), internal cross-references (example_),
-external hyperlinks with embedded URIs (`Python web site <http://www.python.org>`__), footnote references
-(manually numbered [1]_, anonymous auto-numbered [#]_, labeled auto-numbered [#label]_, or symbolic [*]_),
-citation references ([12]_), substitution references (|example|), and _`inline hyperlink targets`
-(see Targets_ below for a reference back to here). Character-level inline markup is also possible
-(although exceedingly ugly!) in *re*\ ``Structured``\ *Text*. Problems are indicated by |problematic|
-text (generated by processing errors; this one is intentional).
-
-Also with ``sphinx.ext.autodoc``, which I use in the demo, I can link to :class:`test_py_module.test.Foo`.
-It will link you right to my code documentation for it.
-
-The default role for [1]_ interpreted text is `Title Reference`.  Here are some explicit interpreted text roles:
-a PEP reference (:PEP:`287`); an RFC reference (:RFC:`2822`); a :sub:`subscript`; a :sup:`superscript`;
-and explicit roles for :emphasis:`standard` :strong:`inline` :literal:`markup`.
-
-GUI labels are a useful way to indicate that :guilabel:`Some action` is to be taken by the user.
-The GUI label should not run over ``line-height`` so as not to :guilabel:`interfere` with text from adjacent lines.
-
-Key-bindings indicate that [1]_ the read is to press a button on the keyboard or mouse,
-for example :kbd:`MMB` and :kbd:`Shift-MMB`. Another useful markup to indicate a user action
-is to use ``menuselection`` this [1]_ can [1]_ be [1]_ used [1]_ to [1]_ show [1]_ short [1]_ and long menus in software.
-For example, and ``menuselection`` can be seen here that breaks is too long to fit on this line.
-:menuselection:`My --> Software --> Some menu --> Some sub menu 1 --> sub menu 2`.
-
-.. DO NOT RE-WRAP THE FOLLOWING PARAGRAPH!
-
-Let's test wrapping and whitespace significance in inline literals:
-``This is an example of --inline-literal --text, --including some--
-strangely--hyphenated-words.  Adjust-the-width-of-your-browser-window
-to see how the text is wrapped.  -- ---- --------  Now note    the
-spacing    between the    words of    this sentence    (words
-should    be grouped    in pairs).``
-
-If the ``--pep-references`` option was supplied, there should be a live link to PEP 258 here.
-
-.. regression test for https://github.com/readthedocs/sphinx_rtd_theme/pull/1193
-
-Very long URLs should be wrapped so lines do not overflow and cause horizontal scrolling: https://www.google.com/search?hl=en&q=very%20long%20url%20example%20of%20a%20url%20that%20is%20extremely%20long%20you%20probably%20want%20to%20avoid%20it%20but%20here%20we%20are
-
-Math
-====
-
-This is a test. Here is an equation:
-:math:`X_{0:5} = (X_0, X_1, X_2, X_3, X_4)`.
-Here is another:
-
-.. math::
-    :label: This is a label
-
-    \nabla^2 f =
-    \frac{1}{r^2} \frac{\partial}{\partial r}
-    \left( r^2 \frac{\partial f}{\partial r} \right) +
-    \frac{1}{r^2 \sin \theta} \frac{\partial f}{\partial \theta}
-    \left( \sin \theta \, \frac{\partial f}{\partial \theta} \right) +
[…]
