### 2023-11-03 caa14ef
Update index.rst with community guidelines
