### 2023-11-03 b5f38c7
Update literature_simio.rst
