+++ paper/paper.bib
@@ -31,6 +31,27 @@ archivePrefix = {arXiv},
       adsnote = {Provided by the SAO/NASA Astrophysics Data System}
 }
 
+@ARTICLE{CASA2022,
+       author = {{CASA Team} and {Bean}, Ben and {Bhatnagar}, Sanjay and {Castro}, Sandra and {Donovan Meyer}, Jennifer and {Emonts}, Bjorn and {Garcia}, Enrique and {Garwood}, Robert and {Golap}, Kumar and {Gonzalez Villalba}, Justo and {Harris}, Pamela and {Hayashi}, Yohei and {Hoskins}, Josh and {Hsieh}, Mingyu and {Jagannathan}, Preshanth and {Kawasaki}, Wataru and {Keimpema}, Aard and {Kettenis}, Mark and {Lopez}, Jorge and {Marvil}, Joshua and {Masters}, Joseph and {McNichols}, Andrew and {Mehringer}, David and {Miel}, Renaud and {Moellenbrock}, George and {Montesino}, Federico and {Nakazato}, Takeshi and {Ott}, Juergen and {Petry}, Dirk and {Pokorny}, Martin and {Raba}, Ryan and {Rau}, Urvashi and {Schiebel}, Darrell and {Schweighart}, Neal and {Sekhar}, Srikrishna and {Shimada}, Kazuhiko and {Small}, Des and {Steeb}, Jan-Willem and {Sugimoto}, Kanako and {Suoranta}, Ville and {Tsutsumi}, Takahiro and {van Bemmel}, Ilse M. and {Verkouter}, Marjolein and {Wells}, Akeem and {Xiong}, Wei and {Szomoru}, Arpad and {Griffith}, Morgan and {Glendenning}, Brian and {Kern}, Jeff},
+        title = "{CASA, the Common Astronomy Software Applications for Radio Astronomy}",
+      journal = {\pasp},
+     keywords = {Single-dish antennas, Aperture synthesis, Radio astronomy, Radio interferometry, Long baseline interferometry, Astronomy software, Open source software, Software documentation, Astronomy data reduction, Astronomy data analysis, 1460, 53, 1338, 1346, 932, 1855, 1866, 1869, 1861, 1858, Astrophysics - Instrumentation and Methods for Astrophysics, Astrophysics - Astrophysics of Galaxies, Astrophysics - High Energy Astrophysical Phenomena, Astrophysics - Solar and Stellar Astrophysics},
+         year = 2022,
+        month = nov,
+       volume = {134},
+       number = {1041},
+          eid = {114501},
+        pages = {114501},
+          doi = {10.1088/1538-3873/ac9642},
+archivePrefix = {arXiv},
+       eprint = {2210.02276},
+ primaryClass = {astro-ph.IM},
+       adsurl = {https://ui.adsabs.harvard.edu/abs/2022PASP..134k4501C},
+      adsnote = {Provided by the SAO/NASA Astrophysics Data System}
+}
+
+
+
 @ARTICLE{Huang:2018,
        author = {{Huang}, Jane and {Andrews}, Sean M. and {Cleeves}, L. Ilsedore and {{\"O}berg}, Karin I. and {Wilner}, David J. and {Bai}, Xuening and {Birnstiel}, Til and {Carpenter}, John and {Hughes}, A. Meredith and {Isella}, Andrea and {P{\'e}rez}, Laura M. and {Ricci}, Luca and {Zhu}, Zhaohuan},
         title = "{CO and Dust Properties in the TW Hya Disk from High-resolution ALMA Observations}",
