### 2023-11-03 893de17
Updated template instructions
