+++ paper/paper.bib
@@ -31,7 +31,7 @@ archivePrefix = {arXiv},
       adsnote = {Provided by the SAO/NASA Astrophysics Data System}
 }
 
-@ARTICLE{CASA2022,
+@ARTICLE{CASA:2022,
        author = {{CASA Team} and {Bean}, Ben and {Bhatnagar}, Sanjay and {Castro}, Sandra and {Donovan Meyer}, Jennifer and {Emonts}, Bjorn and {Garcia}, Enrique and {Garwood}, Robert and {Golap}, Kumar and {Gonzalez Villalba}, Justo and {Harris}, Pamela and {Hayashi}, Yohei and {Hoskins}, Josh and {Hsieh}, Mingyu and {Jagannathan}, Preshanth and {Kawasaki}, Wataru and {Keimpema}, Aard and {Kettenis}, Mark and {Lopez}, Jorge and {Marvil}, Joshua and {Masters}, Joseph and {McNichols}, Andrew and {Mehringer}, David and {Miel}, Renaud and {Moellenbrock}, George and {Montesino}, Federico and {Nakazato}, Takeshi and {Ott}, Juergen and {Petry}, Dirk and {Pokorny}, Martin and {Raba}, Ryan and {Rau}, Urvashi and {Schiebel}, Darrell and {Schweighart}, Neal and {Sekhar}, Srikrishna and {Shimada}, Kazuhiko and {Small}, Des and {Steeb}, Jan-Willem and {Sugimoto}, Kanako and {Suoranta}, Ville and {Tsutsumi}, Takahiro and {van Bemmel}, Ilse M. and {Verkouter}, Marjolein and {Wells}, Akeem and {Xiong}, Wei and {Szomoru}, Arpad and {Griffith}, Morgan and {Glendenning}, Brian and {Kern}, Jeff},
         title = "{CASA, the Common Astronomy Software Applications for Radio Astronomy}",
       journal = {\pasp},
