+++ paper/paper.md
@@ -29,12 +29,12 @@ The documentation and tutorials give a detailed description of the code function
 
 Interferometric facilities, such as ALMA, allow us to reach angular resolutions inaccessible by single-dish telescopes. A crucial step in understanding interferometric datasets is generating synthetic observations to compare or predict how physical processes or intensity distributions would be observed. 
 
-Predicting tools from the Common Astronomy Software Applications package (\texttt{CASA}, [@McMullin:2007, @Emonts:2020, @CASA2022]) are very well suited for generating synthetic observations with specific observational setups (Antenna Array configuration, atmospheric conditions, sky coordinates, exposure time, among others), but generating synthetic observations to mimic existing datasets can be challenging due to the high number of parameters that need to be matched. The SIMIO-continuum package contains all the necessary tools to generate synthetic observations to mimic existing observations without the need to fit every observational parameter manually. 
+Predicting tools from the Common Astronomy Software Applications package (\texttt{CASA}, [@McMullin:2007, @Emonts:2020, @CASA:2022]) are very well suited for generating synthetic observations with specific observational setups (Antenna Array configuration, atmospheric conditions, sky coordinates, exposure time, among others), but generating synthetic observations to mimic existing datasets can be challenging due to the high number of parameters that need to be matched. The SIMIO-continuum package contains all the necessary tools to generate synthetic observations to mimic existing observations without the need to fit every observational parameter manually. 
 
 
 # State of the field
 
-SIMIO-continuum is a python-based package for \texttt{CASA}, designed to take simulated images of millimeter continuum emission and return a synthetic interferometric observation and its reconstructed images. This functionality is particularly useful when comparing simulated models to archival observations. A common way to do this is by convolving the model image with a Gaussian representative of the angular resolution of the observation. Although this method is quick and easy to apply, it only considers a small step of the imaging reconstruction algorithms. Convolving with a gaussian does not take into account other factors that could impact the brightness distribution of a reconstructed image [see introduction section of @Czekala:2021 and references therein]. On the other hand, the \texttt{CASA} task \texttt{simobserve} allows to create a new observational setup to include the simulated image [e.g., @Barraza:2021], but this requires a level of knowledge of interferometric observations. Additionally, it can become tedious if the goal is to reproduce an existing observational setup composed of several projects [e.g., @Huang:2018].
+SIMIO-continuum is a python-based package for \texttt{CASA}, designed to take simulated images of millimeter continuum emission and return a synthetic interferometric observation and its reconstructed images. This functionality is particularly useful when comparing simulated models to archival observations. A common way to do this is by convolving the model image with a Gaussian representative of the angular resolution of the observation. Although this method is quick and easy to apply, it only considers a small step of the imaging reconstruction algorithms. Convolving with a Gaussian does not take into account other factors that could impact the brightness distribution of a reconstructed image [see introduction section of @Czekala:2021 and references therein]. On the other hand, the \texttt{CASA} task \texttt{simobserve} allows to create a new observational setup to include the simulated image [e.g., @Barraza:2021], but this requires a level of knowledge of interferometric observations. Additionally, it can become tedious if the goal is to reproduce an existing observational setup composed of several projects [e.g., @Huang:2018].
 
 Instead of creating a new observational setup as \texttt{simobserve}, SIMIO-continuum uses an existing observation as a template for the synthetic observation. The package takes an input image (containing a millimeter continuum brightness distribution), calculates its visibilities, and then replaces this data in the template observation. This way, the synthetic observation mimics all the technical details of the template (such as the number of antennas, exposure time, frequency coverage, sky orientation, time of observation, and angular resolution). Once the synthetic observation is created, the synthetic reconstructed images are obtained the same way as real observations.
 
