+++ README.md
@@ -30,7 +30,7 @@ How would the Solar System have looked if it was located where the HD163296 syst
 Click here to download this example project: [SolarS_HD163296.tar.gz](https://keeper.mpdl.mpg.de/f/23c92e46e69c40e185f1/) (163.6 MB). After downloading it, untar it in the **projects** folder, such that the directories have the following order: /path_to_SIMIO/SIMIO/projects/SolarS_HD163296. 
 The template for this project is the DSHARP observation of HD163296, which you can download from this [link](https://www.nicolaskurtovic.com/simio/templates). After downloading it, untar it in the **templates** folder, such that the directories have the following order: /path_to_SIMIO/SIMIO/templates/HD163296. 
 
-Follow the [tutorial 1](https://simio-continuum.readthedocs.io) to go step by step from model to synthetic observation.
+Follow the [tutorial 1](https://simio-continuum.readthedocs.io/index.html#document-tutorials/tutorial_1) to go step by step from model to synthetic observation.
 
 <p align="center">
 <img src="https://github.com/nicokurtovic/SIMIO/blob/main/docs/_static/SolarS_HD163296_comparison.png" width="1024"/>
