### 2023-11-06 11357c0
Updated to work with CASA version >6
Changed the execution of external python codes
