### 2023-11-06 130b6f8
Updated JvM correction to work with CASA >6
