### 2023-11-06 2b64323
Updated functions for geometry change
