### 2023-11-06 991f32a
Updated file execution for CASA >6
