+++ docs/index.rst
@@ -26,7 +26,7 @@ indistinctively.
    
    content/api_simio
    content/simio_installation
-   tutorials/testsimio_0
+   content/testsimio_0
    content/considerations
    content/literature_simio
    content/community_guidelines
