### 2023-11-07 nicokurtovic/SIMIO#9 @nicokurtovic
Changes have been implemented in all the codes, and SIMIO-continuum has been updated to work in CASA 5.X and at least in CASA 6.3. The changes in the code are mostly related to going from python2 to python3. A minor change was implemented to make the JvM correction work in v>6. All the changes are internal to the codes, and no changes have been done to the way in which the user uses SIMIO.

I will maintain the recommendation of running it in CASA v<6, where I have tested it the most, but there should not be problems with running in v>6.
