### 2023-11-07 nicokurtovic/SIMIO#4 @nicokurtovic
Thank you very much for bringing this reference to my attention. I have included it in the bibliography of paper.md
