### 2023-11-07 nicokurtovic/SIMIO#5 @nicokurtovic
Thanks!
The links have been updated to point directly to the mentioned content, instead of the general documentation. I have implemented this changes in all the contents and tutorials.
