### 2023-11-07 nicokurtovic/SIMIO#8 @nicokurtovic
I have included a note mentioning that the product images are already contained in the project folder that can be downloaded for the tutorial. I also clarified that running SIMIO will overwrite any existing image in the project.
