### 2023-11-07 nicokurtovic/SIMIO#3 @nicokurtovic
Thank you very much for your suggestions.

One of the main ideas of the package is to be as simple to use as possible. In its current version, SIMIO does not need any additional dependency to be installed, neither internal installs of the package. The package is ready to use after being downloaded, as long as a template and a project have been included. I moved the first and second step of tutorial 1 to the installation.

Since the package is executed from the CASA terminal, the python libraries such as numpy or scipy are already included in the python of CASA, and thus the user does not need to install them independently in their own python. I included additional clarification that CASA needs to be installed in order to use SIMIO.
