### 2023-11-07 nicokurtovic/SIMIO#6 @nicokurtovic
Thank you for noticing that some information was missing. I have included in the README.md further details of where to get the template, and where to put it after downloading.
