+++ paper/paper.md
@@ -21,7 +21,7 @@ bibliography: paper.bib
 
 # Summary
 
-Interferometric observations sample the visibility space of the targeted source. Thus, procedures for image reconstruction are needed to obtain images from those observations. Several factors are involved in the detectability of physical features in a reconstructed image, such as the angular resolution of the observation, point spread function, noise level, and the assumptions of the reconstruction algorithm. The most robust way to test the observability of simulated emission features is by taking a simulated image and reconstructing it with the same algorithms. SIMIO-continuum takes simulated images of continuum emission in millimeter wavelengths and returns the synthetic observation and reconstructed images as if the input image had been observed in the sky with an existing interferometric observation.
+Interferometric observations sample the visibility space of the targeted source. Thus, procedures for image reconstruction are needed to obtain the sky brightness distribution from those observations. Several factors are involved in the detectability of physical features in a reconstructed image, such as the angular resolution of the observation, point spread function, noise level, and the assumptions of the reconstruction algorithm. The most robust way to test the observability of simulated emission features is by taking a simulated image and reconstructing it with the same algorithms. SIMIO-continuum takes simulated images of continuum emission in millimeter wavelengths and returns the synthetic observation and reconstructed images as if the input image had been observed in the sky with an existing interferometric observation.
 
 The documentation and tutorials give a detailed description of the code functionalities and syntax, with publicly available datasets for reproducibility.
 
@@ -29,7 +29,7 @@ The documentation and tutorials give a detailed description of the code function
 
 Interferometric facilities, such as ALMA, allow us to reach angular resolutions inaccessible by single-dish telescopes. A crucial step in understanding interferometric datasets is generating synthetic observations to compare or predict how physical processes or intensity distributions would be observed. 
 
-Predicting tools from the Common Astronomy Software Applications package (\texttt{CASA}, [@McMullin:2007, @Emonts:2020, @CASA:2022]) are very well suited for generating synthetic observations with specific observational setups (Antenna Array configuration, atmospheric conditions, sky coordinates, exposure time, among others), but generating synthetic observations to mimic existing datasets can be challenging due to the high number of parameters that need to be matched. The SIMIO-continuum package contains all the necessary tools to generate synthetic observations to mimic existing observations without the need to fit every observational parameter manually. 
+Predicting tools from the Common Astronomy Software Applications package (\texttt{CASA}, [@McMullin:2007, @Emonts:2020, @CASA:2022]) are very well suited for generating synthetic observations with specific observational setups (Antenna Array configuration, atmospheric conditions, sky coordinates, exposure time, among others), but generating synthetic observations to mimic existing datasets can be challenging due to the high number of parameters that need to be matched. The SIMIO-continuum package contains all the necessary tools to generate synthetic observations to mimic existing observations without the need to fit every observational parameter manually, going from model image to reconstructed image without the need of previous interferometric knowledge.
 
 
 # State of the field
@@ -37,6 +37,12 @@ Predicting tools from the Common Astronomy Software Applications package (\textt
[…]
