+++ paper/paper.bib
@@ -107,3 +107,31 @@ archivePrefix = {arXiv},
        adsurl = {https://ui.adsabs.harvard.edu/abs/2021ApJS..257....2C},
       adsnote = {Provided by the SAO/NASA Astrophysics Data System}
 }
+
+@ARTICLE{Tazzari:2018,
+       author = {{Tazzari}, Marco and {Beaujean}, Frederik and {Testi}, Leonardo},
+        title = "{GALARIO: a GPU accelerated library for analysing radio interferometer observations}",
+      journal = {\mnras},
+     keywords = {methods: numerical, techniques: interferometric, submillimetre: general, Astrophysics - Instrumentation and Methods for Astrophysics, Astrophysics - Cosmology and Nongalactic Astrophysics, Astrophysics - Earth and Planetary Astrophysics, Astrophysics - Astrophysics of Galaxies, Astrophysics - Solar and Stellar Astrophysics},
+         year = 2018,
+        month = jun,
+       volume = {476},
+       number = {4},
+        pages = {4527-4542},
+          doi = {10.1093/mnras/sty409},
+archivePrefix = {arXiv},
+       eprint = {1709.06999},
+ primaryClass = {astro-ph.IM},
+       adsurl = {https://ui.adsabs.harvard.edu/abs/2018MNRAS.476.4527T},
+      adsnote = {Provided by the SAO/NASA Astrophysics Data System}
+}
+
+@misc{Olberg:2021,
+  author = {{Olberg}, Michael},
+  title = {fakeobs},
+  year = {2021},
+  publisher = {GitHub},
+  journal = {GitHub repository},
+  howpublished = {\url{https://github.com/onsala-space-observatory/fakeobs}},
+  commit = {e591e85}
+}
