+++ paper/paper.md
@@ -29,7 +29,7 @@ The documentation and tutorials give a detailed description of the code function
 
 Interferometric facilities, such as ALMA, allow us to reach angular resolutions inaccessible by single-dish telescopes. A crucial step in understanding interferometric datasets is generating synthetic observations to compare or predict how physical processes or intensity distributions would be observed. 
 
-Predicting tools from the Common Astronomy Software Applications package (\texttt{CASA}, [@McMullin:2007, @Emonts:2020, @CASA:2022]) are very well suited for generating synthetic observations with specific observational setups (Antenna Array configuration, atmospheric conditions, sky coordinates, exposure time, among others), but generating synthetic observations to mimic existing datasets can be challenging due to the high number of parameters that need to be matched. The SIMIO-continuum package contains all the necessary tools to generate synthetic observations to mimic existing observations without the need to fit every observational parameter manually, going from model image to reconstructed image without the need of previous interferometric knowledge.
+Predicting tools from the Common Astronomy Software Applications package (CASA, [@McMullin:2007], [@Emonts:2020], [@CASA:2022]) are very well suited for generating synthetic observations with specific observational setups (Antenna Array configuration, atmospheric conditions, sky coordinates, exposure time, among others), but generating synthetic observations to mimic existing datasets can be challenging due to the high number of parameters that need to be matched. The SIMIO-continuum package contains all the necessary tools to generate synthetic observations to mimic existing observations without the need to fit every observational parameter manually, going from model image to reconstructed image without the need of previous interferometric knowledge.
 
 
 # State of the field
@@ -38,11 +38,7 @@ SIMIO-continuum is a python-based package for \texttt{CASA}, designed to take si
 
 Instead of creating a new observational setup as \texttt{simobserve}, SIMIO-continuum uses an existing observation as a template for the synthetic observation. The package takes an input image (containing a millimeter continuum brightness distribution), calculates its visibilities, and then replaces this data in the template observation. This way, the synthetic observation mimics all the technical details of the template (such as the number of antennas, exposure time, frequency coverage, sky orientation, time of observation, and angular resolution). Once the synthetic observation is created, the synthetic reconstructed images are obtained the same way as real observations.
 
-Additional to simobserve, there are alternative tools to obtain the visibilities from a model image. For example, the package \texttt{galario} [@Tazzari:2018]
-
-github citation [e.g., @Olberg:2021]
-
-
+Additional to simobserve, there are other public tools to obtain the visibilities from a model image. For example, the package \texttt{galario} [@Tazzari:2018] allows to obtain the visibilities of a model image in a python environment, and it is optimized to be used tih Markov Chain Monte Carlo (MCMC) algorithms. Another similar tool is the task \texttt{fakeobs} [fakeobs:2021], which can copy the structure of an existing measurement set and replace the data with the visibilities of a model image. SIMIO-continuum is not meant as a visibility fitting tool, as \texttt{galario}, but rather as a tool to generate synthetic observations. Similarly to \texttt{fakeobs}, SIMIO-continuum also mimic the structure of a measurement set and replaces the data with the visibilities of a model image, but it also includes the option to adjust a model inclination, rotation, distance, and generate CLEAN images with minimum user input. 
 
 SIMIO-continuum was designed to be easy to use by non-observers while offering the full range of data products for people with different observational expertise. Based on a template observation, astronomers with little or no interferometric experience can obtain synthetic observations and images of their models. Additionally, SIMIO-continuum allows changing the model's geometry, including white noise in the reconstructed images, and changing the model distance to the observer, thus enabling multiple feature-recovery tests with a visibility-based approach.
 
