+++ paper/paper.md
@@ -23,22 +23,22 @@ bibliography: paper.bib
 
 Interferometric observations sample the visibility space of the targeted source. Thus, procedures for image reconstruction are needed to obtain the sky brightness distribution from those observations. Several factors are involved in the detectability of physical features in a reconstructed image, such as the angular resolution of the observation, point spread function, noise level, and the assumptions of the reconstruction algorithm. The most robust way to test the observability of simulated emission features is by taking a simulated model image and reconstructing it with the same algorithms. SIMIO-continuum takes simulated images of continuum emission in millimeter wavelengths and returns the synthetic observation and reconstructed images as if the input model had been observed in the sky with an existing interferometric observation.
 
-The documentation and tutorials give a detailed description of the code functionalities and syntax, with publicly available datasets for reproducibility.
+The documentation and tutorials give a detailed description of the code functionalities and syntax with publicly available datasets for reproducibility.
 
 # Statement of need
 
-Interferometric facilities, such as ALMA, allow us to reach angular resolutions inaccessible by single-dish telescopes. A crucial step in understanding interferometric datasets is generating synthetic observations to compare or predict how physical processes or intensity distributions would be observed. 
+Interferometric facilities, such as the Atacama Large sub-Millimeter Array (ALMA), allow us to reach angular resolutions inaccessible by single-dish telescopes. A crucial step in understanding interferometric datasets is generating synthetic observations to compare or predict how physical processes or intensity distributions would be observed. 
 
-The Common Astronomy Software Applications package (CASA, [@McMullin:2007], [@Emonts:2020], [@CASA:2022]) has the necessary tasks to generate synthetic observations and reconstruct images, and they are very well suited for generating very specific observational setups (Antenna Array configuration, atmospheric conditions, sky coordinates, exposure time, among others). However, using these tools requires some interferometric background knowledge, and generating synthetic observations to mimic existing datasets can be challenging due to the high number of parameters that need to be matched. The SIMIO-continuum package contains all the necessary tools to generate a synthetic observation, modify th geometry and distance of a model, and then reconstruct synthetic images as if they were a real interferometric observation. 
+The Common Astronomy Software Applications package (CASA, [@McMullin:2007], [@Emonts:2020], [@CASA:2022]) has the necessary tasks to generate synthetic observations and reconstruct images, and they are very well suited for generating very specific observational setups (Antenna Array configuration, atmospheric conditions, sky coordinates, exposure time, among others). However, using these tools requires some interferometric background knowledge, and generating synthetic observations to mimic existing datasets can be challenging due to the high number of parameters that need to be matched. The SIMIO-continuum package contains all the necessary tools to generate a synthetic observation, modify the geometry and distance of a model, and then reconstruct synthetic images as if they were actual interferometric observations. 
 
 
 # State of the field
 
-SIMIO-continuum is a python-based package for \texttt{CASA}, designed to take simulated images of millimeter continuum emission and return a synthetic interferometric observation and its reconstructed images. This functionality is particularly useful when comparing simulated models to archival observations. A common way to do this is by convolving the model image with a Gaussian representative of the angular resolution of the observation. Although this method is quick and easy to apply, it only considers a small step of the imaging reconstruction algorithms. Convolving with a Gaussian does not take into account other factors that could impact the brightness distribution of a reconstructed image [see introduction section of @Czekala:2021 and references therein]. On the other hand, the \texttt{CASA} task \texttt{simobserve} allows to create a new observational setup to include the simulated image [e.g., @Barraza:2021], but the complexity of this approach can quickly increase if the goal is to reproduce an existing observational setup composed of several projects [e.g., @Huang:2018].
[…]
