### 2023-11-08 nicokurtovic/SIMIO#14 @nicokurtovic
I am very sorry I did not mention that the projection and deprojection of the visibilities was initially based on a frank function. I modified it some time ago to include the shifting in space, and it was my mistake to forget where it originally came from. I should have mentioned it, as I also mentioned the origin of other functions.

I changed the structure of the deprojection function. The shifting, rotation, projection, and deprojection is done in separate tasks, which are all called from the [handle_uv](https://github.com/nicokurtovic/SIMIO/blob/main/codes/simio_support.py#L117) function.

I decided to exclude galario and the functions calling galario from SIMIO. Using galario was only optional in the package, and now the visibilities are only calculated with the Fourier Transform of CASA.
