### 2023-11-08 nicokurtovic/SIMIO#11 @nicokurtovic
Thank you very much for your suggestions. I have modified paper.md to mention these tools and include their references.
