+++ paper/paper.md
@@ -29,7 +29,7 @@ The documentation and tutorials give a detailed description of the code function
 
 Interferometric facilities, such as the Atacama Large sub-Millimeter Array (ALMA), allow us to reach angular resolutions inaccessible by single-dish telescopes. A crucial step in understanding interferometric datasets is generating synthetic observations to compare or predict how physical processes or intensity distributions would be observed. 
 
-The Common Astronomy Software Applications package (CASA, [@McMullin:2007], [@Emonts:2020], [@CASA:2022]) has the necessary tasks to generate synthetic observations and reconstruct images, and they are very well suited for generating very specific observational setups (Antenna Array configuration, atmospheric conditions, sky coordinates, exposure time, among others). However, using these tools requires some interferometric background knowledge, and generating synthetic observations to mimic existing datasets can be challenging due to the high number of parameters that need to be matched. The SIMIO-continuum package contains all the necessary tools to generate a synthetic observation, modify the geometry and distance of a model, and then reconstruct synthetic images as if they were actual interferometric observations. 
+The Common Astronomy Software Applications package [CASA;@McMullin:2007;@Emonts:2020;@CASA:2022] has the necessary tasks to generate synthetic observations and reconstruct images, and they are very well suited for generating very specific observational setups (Antenna Array configuration, atmospheric conditions, sky coordinates, exposure time, among others). However, using these tools requires some interferometric background knowledge, and generating synthetic observations to mimic existing datasets can be challenging due to the high number of parameters that need to be matched. The SIMIO-continuum package contains all the necessary tools to generate a synthetic observation, modify the geometry and distance of a model, and then reconstruct synthetic images as if they were actual interferometric observations. 
 
 
 # State of the field
