### 2023-11-21 0e91404
Updated text and references
