+++ paper/paper.md
@@ -14,7 +14,7 @@ authors:
 affiliations:
  - name: Max-Planck-Institut f\"{u}r Astronomie, K\"{o}nigstuhl 17, 69117, Heidelberg, Germany.
    index: 1
-date: 8 November 2023
+date: 21 November 2023
 bibliography: paper.bib
 
 ---
@@ -29,18 +29,18 @@ The documentation and tutorials give a detailed description of the code function
 
 Interferometric facilities, such as the Atacama Large sub-Millimeter Array (ALMA), allow us to reach angular resolutions inaccessible by single-dish telescopes. A crucial step in understanding interferometric datasets is generating synthetic observations to compare or predict how physical processes or intensity distributions would be observed. 
 
-The Common Astronomy Software Applications package [CASA;@McMullin:2007;@Emonts:2020;@CASA:2022] has the necessary tasks to generate synthetic observations and reconstruct images, and they are very well suited for generating very specific observational setups (Antenna Array configuration, atmospheric conditions, sky coordinates, exposure time, among others). However, using these tools requires some interferometric background knowledge, and generating synthetic observations to mimic existing datasets can be challenging due to the high number of parameters that need to be matched. The SIMIO-continuum package contains all the necessary tools to generate a synthetic observation, modify the geometry and distance of a model, and then reconstruct synthetic images as if they were actual interferometric observations. 
+The Common Astronomy Software Applications package (CASA; @McMullin:2007; @Emonts:2020; @CASA:2022) has the necessary tasks to generate synthetic observations and reconstruct images, and they are very well suited for generating very specific observational setups (Antenna Array configuration, atmospheric conditions, sky coordinates, exposure time, among others). However, generating synthetic observations to mimic existing datasets can be challenging due to the high number of parameters that need to be matched. The SIMIO-continuum package contains all the necessary tools to simplify the process of generating synthetic observations, including routines to modify the geometry and relative distance of a model, and reconstructing synthetic images as if they were actual interferometric observations. 
 
 
 # State of the field
 
-SIMIO-continuum is a python-based package for \texttt{CASA}, designed to take simulated images of millimeter continuum emission and return a synthetic interferometric observation and its reconstructed images. This functionality is particularly useful when comparing simulated models to archival observations. A common way to do this is by convolving the model image with a Gaussian representative of the angular resolution of the observation. Although this method is quick and easy to apply, it only considers a small step of the imaging reconstruction algorithms. Convolving with a Gaussian does not take into account other factors that could impact the brightness distribution of a reconstructed image [see introduction section of @Czekala:2021 and references therein]. On the other hand, the \texttt{CASA} task \texttt{simobserve} creates a new observational setup to include the simulated image [e.g., @Barraza:2021], but the complexity of this approach can quickly increase if the goal is to reproduce an existing observational setup composed of several projects [e.g., @Huang:2018].
+SIMIO-continuum is a python-based package for \texttt{CASA}, designed to take simulated images of millimeter continuum emission and return a synthetic interferometric observation and its reconstructed images. This functionality is particularly useful when comparing simulated models to archival observations. A common way to do this is by convolving the model image with a Gaussian representative of the angular resolution of the observation. Although this method is quick and easy to apply, it only considers a small step of the imaging reconstruction algorithms. Convolving with a Gaussian does not take into account other factors that could impact the brightness distribution of a reconstructed image (see introduction section of @Czekala:2021 and references therein). On the other hand, the \texttt{CASA} task \texttt{simobserve} creates a new observational setup to include the simulated image [e.g., @Barraza:2021], but the complexity of this approach can quickly increase if the goal is to reproduce an existing observational setup composed of several projects [e.g., @Huang:2018].
 
[…]
