### 2023-11-21 nicokurtovic/SIMIO#13 @nicokurtovic
Hi Jeff, thanks for the suggestion.

I have been thinking a lot about this, and at least for this version I would prefer to keep it as an executable. The initial inspiration for the code structure was the hydro-code [FARGO 3D](https://fargo3d.bitbucket.io/index.html), where you work inside the FARGO folder and you can save your setups individually. My thought was to have something similar to FARGO, where you would only need to download SIMIO, and then save the templates and projects individually in a self-contained folder.

I think since the last CASA 6.X versions, CASA tasks can also be called from the python terminal ([example](https://casadocs.readthedocs.io/en/stable/notebooks/usingcasa.html)), and so a future SIMIO package could be fully integrated in python, but for the current SIMIO I would prefer to keep the folder structure, if possible.
