### 2023-11-21 nicokurtovic/SIMIO#10 @nicokurtovic
Thank you for noticing this typos, they should be fixed in the current version
