### 2023-11-21 nicokurtovic/SIMIO#2 @nicokurtovic
A statement of need has been included in the new version of the JOSS paper
