+++ docs/index.rst
@@ -6,7 +6,7 @@ your radiative transfer images with existing ALMA observations. With
 *SIMIO-continuum* you can generate your synthetic observations with
 different geometries, distances, and even including white noise.
 
-Please check www.nicolaskurtovic.com/simio and the Contents section
+Please check `www.nicolaskurtovic.com/simio <https://www.nicolaskurtovic.com/simio>`_ and the Contents section
 for a detailed description of the package. Although *SIMIO-continuum*
 is the official name, this page uses *SIMIO-continuum* or *SIMIO*
 indistinctively.
