+++ codes/simio_support.py
@@ -25,7 +25,7 @@ import numpy as np
 
 def _shift_uv(u, v, Re, Im, dRa, dDec):
     '''
-    Applies rotation and inclination to the input visibilities.
+    Applies spatial shift to the visibilities
     
     Args:
         - u: (numpy array) Array containing the ``u`` coordinates of the
@@ -36,8 +36,8 @@ def _shift_uv(u, v, Re, Im, dRa, dDec):
                     visibilities.
         - Im: (numpy array) Array containing the ``Imaginary`` part of the
                     visibilities.
-        - inc: (float) Value to incline the visibilities, in **degrees**.
-        - pa: (float) Value to rotate the visibilities, in **degrees**.
+        - dRa: (float) Value to shift the visiblities in RA, in **radians**.
+        - dDec: (float) Value to shift the visiblities in Dec, in **radians**.
     '''
     # Calculate the value of the visibilities
     intensity = (Re + Im * 1j) * np.exp(2j * np.pi * (u * -dRa + v * -dDec))
@@ -72,19 +72,14 @@ def _rotate_uv(u, v, pa):
 
 def _deproject_uv(u, v, inc):
     '''
-    Applies rotation and inclination to the input visibilities.
+    Applies an inverse inclination to the input visibilities.
     
     Args:
         - u: (numpy array) Array containing the ``u`` coordinates of the
                     visibilities.
         - v: (numpy array) Array containing the ``v`` coordinates of the
                     visibilities.
-        - Re: (numpy array) Array containing the ``Real`` part of the
-                    visibilities.
-        - Im: (numpy array) Array containing the ``Imaginary`` part of the
-                    visibilities.
-        - inc: (float) Value to incline the visibilities, in **degrees**.
-        - pa: (float) Value to rotate the visibilities, in **degrees**.
+        - inc: (float) Value to incline the visibilities, in **radians**.
     '''
     # Apply inclination
     u_inc = u * np.cos(inc)
@@ -94,19 +89,14 @@ def _deproject_uv(u, v, inc):
 
 def _project_uv(u, v, inc):
     '''
-    Applies rotation and inclination to the input visibilities.
+    Applies an inclination to the input visibilities.
     
     Args:
         - u: (numpy array) Array containing the ``u`` coordinates of the
                     visibilities.
         - v: (numpy array) Array containing the ``v`` coordinates of the
                     visibilities.
-        - Re: (numpy array) Array containing the ``Real`` part of the
-                    visibilities.
-        - Im: (numpy array) Array containing the ``Imaginary`` part of the
-                    visibilities.
-        - inc: (float) Value to incline the visibilities, in **degrees**.
-        - pa: (float) Value to rotate the visibilities, in **degrees**.
+        - inc: (float) Value to incline the visibilities, in **radians**.
     '''
     # Apply inclination
     u_inc = u / np.cos(inc)    
@@ -116,8 +106,8 @@ def _project_uv(u, v, inc):
 
 def handle_uv(u, v, Re, Im, inc, pa, dRa, dDec, inverse=False):
     '''
-    Projects, deprojects, and shifts visibility points based on the input
-    parameters.
+    Projects, deprojects, rotates, and shifts visibility points based on the
+    input parameters.
     
     Args:
         - u: (numpy array) Array containing the ``u`` coordinates of the
@@ -145,7 +135,7 @@ def handle_uv(u, v, Re, Im, inc, pa, dRa, dDec, inverse=False):
     if not inverse:
         # Calculate shifted visibilities
         u, v, Re_shift, Im_shift = _shift_uv(u, v, Re, Im, dRa, dDec)
-        # Cprrect rotation
[…]
