### 2023-11-21 67701c3
Update README.md with v1.2 SIMIO
