+++ paper/paper.bib
@@ -126,8 +126,8 @@ archivePrefix = {arXiv},
       adsnote = {Provided by the SAO/NASA Astrophysics Data System}
 }
 
-@misc{Olberg:2021,
-  author = {{Olberg}, Michael},
+@misc{fakeobs:2021,
+  author = {{Nordic ALMA Regional Center} and {Olberg}, Michael},
   title = {fakeobs},
   year = {2021},
   publisher = {GitHub},
