+++ paper/paper.bib
@@ -127,11 +127,11 @@ archivePrefix = {arXiv},
 }
 
 @misc{fakeobs:2021,
-  author = {{Nordic ALMA Regional Center} and {Olberg}, Michael},
-  title = {fakeobs},
-  year = {2021},
-  publisher = {GitHub},
-  journal = {GitHub repository},
-  howpublished = {\url{https://github.com/onsala-space-observatory/fakeobs}},
-  commit = {e591e85}
+      author = {{Nordic ALMA Regional Center} and {Olberg}, Michael},
+      title = {fakeobs},
+      year = {2021},
+      publisher = {GitHub},
+      journal = {GitHub repository},
+      howpublished = {\url{https://github.com/onsala-space-observatory/fakeobs}},
+      commit = {e591e85}
 }
