### 2023-11-21 a4ebcd8
Update paper.bib
updated fakeobs information
