+++ paper/paper.bib
@@ -131,7 +131,9 @@ archivePrefix = {arXiv},
       title = {fakeobs},
       year = {2021},
       publisher = {GitHub},
+      type = {Repository},
       journal = {GitHub repository},
+      url = {https://github.com/onsala-space-observatory/fakeobs}
       howpublished = {\url{https://github.com/onsala-space-observatory/fakeobs}},
       commit = {e591e85}
 }
