### 2023-11-21 6cc9e65
Update paper.bib
updated fakeobs
