+++ paper/paper.bib
@@ -134,6 +134,5 @@ archivePrefix = {arXiv},
       type = {Repository},
       journal = {GitHub repository},
       url = {https://github.com/onsala-space-observatory/fakeobs}
-      howpublished = {\url{https://github.com/onsala-space-observatory/fakeobs}},
       commit = {e591e85}
 }
