+++ paper/paper.bib
@@ -133,6 +133,6 @@ archivePrefix = {arXiv},
       publisher = {GitHub},
       type = {Repository},
       journal = {GitHub repository},
-      url = {https://github.com/onsala-space-observatory/fakeobs}
+      url = {https://github.com/onsala-space-observatory/fakeobs},
       commit = {e591e85}
 }
