### 2023-11-21 #4942 @nicokurtovic
Dear @ivastar , @jeffjennings and @adeleplunkett , I am very sorry for the delay.

I have updated the code, documentation, webpage, and paper, addressing the comments that were raised in the issues section of SIMIO. I left some of the issues open in case you would like to comment on the implemented solution.

Thank you for your time
