### 2023-11-30 nicokurtovic/SIMIO#15 @jeffjennings
Minor typo in paper.md
In l.34 of paper.pdf, "SIMIO-continuum is a python-based package for CASA", capitalize Python.
