### 2023-11-30 nicokurtovic/SIMIO#11 @jeffjennings
I see the updated discussion in the paper, thanks!
