### 2023-11-30 nicokurtovic/SIMIO#7 @jeffjennings
I see the added short guidelines in the [docs](https://simio-continuum.readthedocs.io/index.html#document-content/community_guidelines).
