### 2023-11-30 nicokurtovic/SIMIO#13 @jeffjennings
Ok, I appreciate you thinking about this. I think ultimately it would benefit users and maintainers of SIMIO if the package is fully self-contained, but the design choice is yours. I added a suggestion about making the nature of SIMIO a bit more clear in the paper in my comment in #3. I'm happy to close this issue otherwise.
