### 2023-11-30 #4942 @jeffjennings
Hi @ivastar, I've reviewed the author's changes and completed my review checklist. I've added a couple of comments to outstanding issues in the code repository ([issue 3](https://github.com/nicokurtovic/SIMIO/issues/3) and [issue 15](https://github.com/nicokurtovic/SIMIO/issues/15)). Once those are resolved, I'll follow up here to note I'm ready to accept.
