### 2023-12-07 nicokurtovic/SIMIO#3 @adeleplunkett
I agree with @jeffjennings suggestion to modify the text slightly to specify this.

I think that the https://simio-continuum.readthedocs.io/index.html#document-content/simio_installation section is good for installation.  I see that there is a file called requirements.txt.  This would be a list of dependencies that should be outlined, right?
