### 2023-12-07 #4942 @adeleplunkett
@ivastar apologies for slow reply here. I thought I replied by email while traveling, but maybe it didn't go through.

I looked over all the revisions. It is very close.  Just two points to note here:

_Installation instructions: Is there a clearly-stated list of dependencies?_ --> Pending to address and close https://github.com/nicokurtovic/SIMIO/issues/3

@nicokurtovic can close https://github.com/nicokurtovic/SIMIO/issues/7
