+++ paper/paper.md
@@ -1,5 +1,5 @@
 ---
-title: 'SIMIO-continuum: Connecting simulations to ALMA observations'
+title: 'SIMIO-continuum: Connecting simulations to interferometric observations'
 tags:
   - Python
   - CASA
@@ -12,39 +12,31 @@ authors:
     orcid: 0000-0002-2358-4796
     affiliation: "1" #
 affiliations:
- - name: Max-Planck-Institut f\"{u}r Astronomie, K\"{o}nigstuhl 17, 69117, Heidelberg, Germany.
+ - name: Max-Planck-Institut für Astronomie, Königstuhl 17, 69117, Heidelberg, Germany.
    index: 1
-date: 21 November 2023
+date: 15 August 2022
 bibliography: paper.bib
 
 ---
 
 # Summary
 
-Interferometric observations sample the visibility space of the targeted source. Thus, procedures for image reconstruction are needed to obtain the sky brightness distribution from those observations. Several factors are involved in the detectability of physical features in a reconstructed image, such as the angular resolution of the observation, point spread function, noise level, and the assumptions of the reconstruction algorithm. The most robust way to test the observability of simulated emission features is by taking a simulated model image and reconstructing it with the same algorithms. SIMIO-continuum takes simulated images of continuum emission in millimeter wavelengths and returns the synthetic observation and reconstructed images as if the input model had been observed in the sky with an existing interferometric observation.
+Interferometric observations sample the visibility space of the targeted source. Thus, procedures for image reconstruction are needed to obtain images from those observations. Several factors are involved in the detectability of physical features in a reconstructed image, such as the angular resolution of the observation, point spread function, noise level, and the assumptions of the reconstruction algorithm. The most robust way to test the observability of simulated emission features is by taking a simulated image and reconstructing it with the same algorithms. SIMIO-continuum takes simulated images of continuum emission in millimeter wavelengths and returns the synthetic observation and reconstructed images as if the input image had been observed in the sky with an existing interferometric observation.
 
-The documentation and tutorials give a detailed description of the code functionalities and syntax with publicly available datasets for reproducibility.
+The documentation and tutorials give a detailed description of the code functionalities and syntax, with publicly available datasets for reproducibility.
 
-# Statement of need
+# Background
 
-Interferometric facilities, such as the Atacama Large sub-Millimeter Array (ALMA), allow us to reach angular resolutions inaccessible by single-dish telescopes. A crucial step in understanding interferometric datasets is generating synthetic observations to compare or predict how physical processes or intensity distributions would be observed. 
+SIMIO-continuum is a Python-based package for \texttt{CASA} [Common Astronomy Software Applications, Emonts:2020; @McMullin:2007], designed to take simulated images of millimeter continuum emission and return a synthetic interferometric observation and its reconstructed images. This functionality is particularly useful when comparing simulated models to archival observations. A common way to do this is by convolving the model image with a Gaussian representative of the angular resolution of the observation. Although this method is quick and easy to apply, it only considers a small step of the imaging reconstruction algorithms. Convolving with a gaussian does not take into account other factors that could impact the brightness distribution of a reconstructed image [see introduction section of @Czekala:2021 and references therein]. On the other hand, the \texttt{CASA} task \texttt{simobserve} allows to create a new observational setup to include the simulated image [e.g., @Barraza:2021], but this requires setting every specific detail of the observation to compare. Additionally, it can become tedious if the goal is to reproduce an existing setup composed of several observations [e.g., @Huang:2018].
 
[…]
