### 2024-02-06 602e428
Updated paper.md file to fix typos
