+++ paper/paper.md
@@ -1,5 +1,5 @@
 ---
-title: 'SIMIO-continuum: Connecting simulations to interferometric observations'
+title: 'SIMIO-continuum: Connecting simulations to ALMA observations'
 tags:
   - Python
   - CASA
@@ -12,9 +12,9 @@ authors:
     orcid: 0000-0002-2358-4796
     affiliation: "1" #
 affiliations:
- - name: Max-Planck-Institut für Astronomie, Königstuhl 17, 69117, Heidelberg, Germany.
+ - name: Max-Planck-Institut f\"{u}r Astronomie, K\"{o}nigstuhl 17, 69117, Heidelberg, Germany.
    index: 1
-date: 15 August 2022
+date: 2 November 2023
 bibliography: paper.bib
 
 ---
@@ -25,18 +25,24 @@ Interferometric observations sample the visibility space of the targeted source.
 
 The documentation and tutorials give a detailed description of the code functionalities and syntax, with publicly available datasets for reproducibility.
 
-# Background
+# Statement of need
 
-SIMIO-continuum is a Python-based package for \texttt{CASA} [Common Astronomy Software Applications, Emonts:2020; @McMullin:2007], designed to take simulated images of millimeter continuum emission and return a synthetic interferometric observation and its reconstructed images. This functionality is particularly useful when comparing simulated models to archival observations. A common way to do this is by convolving the model image with a Gaussian representative of the angular resolution of the observation. Although this method is quick and easy to apply, it only considers a small step of the imaging reconstruction algorithms. Convolving with a gaussian does not take into account other factors that could impact the brightness distribution of a reconstructed image [see introduction section of @Czekala:2021 and references therein]. On the other hand, the \texttt{CASA} task \texttt{simobserve} allows to create a new observational setup to include the simulated image [e.g., @Barraza:2021], but this requires setting every specific detail of the observation to compare. Additionally, it can become tedious if the goal is to reproduce an existing setup composed of several observations [e.g., @Huang:2018].
+Interferometric facilities, such as ALMA, allow us to reach angular resolutions inaccessible by single-dish telescopes. A crucial step in understanding interferometric datasets is generating synthetic observations to compare or predict how physical processes or intensity distributions would be observed. 
 
-Instead of creating a new observational setup as \texttt{simobserve}, SIMIO-continuum uses an existing observation as a template for the synthetic observation. The package takes an input image (containing a millimeter continuum brightness distribution), calculates its visibilities, and then replaces this data in the template observation. In this way, the synthetic observation mimics all the technical details of the template (such as the number of antennas, exposure time, frequency coverage, sky orientation, time of observation, and angular resolution). Once the synthetic observation is created, the synthetic reconstructed images are obtained in the same way as real observations.
+Predicting tools from the Common Astronomy Software Applications package (\texttt{CASA}, [@McMullin:2007, @Emonts:2020]) are very well suited for generating synthetic observations with specific observational setups (Antenna Array configuration, atmospheric conditions, sky coordinates, exposure time, among others), but generating synthetic observations to mimic existing datasets can be challenging due to the high number of parameters that need to be matched. The SIMIO-continuum package contains all the necessary tools to generate synthetic observations to mimic existing observations without the need to fit every observational parameter manually. 
 
-SIMIO-continuum was designed to be easy to use by non-experts on interferometric observations while offering a full range of data products for people with different observational backgrounds. Based on a template observation, astronomers can obtain not only the synthetic observations and images of their models, as SIMIO-continuum allows changing the model's geometry[^1], including white noise in the reconstructed images, and changing the model distance to the observer. These features enable multiple feature-recovery tests with a visibility-based approach.
 
[…]
