### 2024-02-07 nicokurtovic/SIMIO#3 @nicokurtovic
Hi Adele,

The list of dependencies in docs/requirements.txt only applies for the webpage in redthedocs and the compilation of the documentation pdf. The reason this list is so long is because I had to test many things to fix the problem with the rendering of the equations.

Other than that, the only requirement for SIMIO-continuum should be CASA v5.X or v6.X, as described in the installation instructions.
