### 2024-02-09 nicokurtovic/SIMIO#3 @jeffjennings
> I see, thank you for considering this, and for adding the clarification. I think it would still be nice to have the package be standalone, but since the current functionality is what's detailed in the paper, of course that's up to you. But could you be a bit more explicit in the paper that SIMIO is a set of programs designed to be run in CASA?
>
> I'd suggest something like changing line 34 in the paper from "SIMIO-continuum is a python-based package for CASA" to "SIMIO-continuum is a set of Python modules designed to be executed within a CASA Python environment".

Thanks for making this change to the wording in the paper. I'm happy for this issue to be closed when Adele is.
