### 2024-03-18 #4942 @ivastar
@nicokurtovic Just one small editorial item:

- On lines 22 and 23 there is a reference that has a wonky format, there should not be brackets around the year, use citep instead of citet.

- I see that it is flagging some missing DOIs, these papers don't seem to have proper DOIs, it boggles the mind. And I don't know what fakeobs is.

I am now going to generate the checkout list. Please follow the steps for generating an archive and getting a DOI for it. Make sure that the names and license on the archive match those on the paper.
