### 2024-03-18 #4942 @ivastar
> Hi @ivastar I'll take a look at the reference with strange format.
>
> There are some citations to codes or packages which do not have a DOI, which were suggested by the referees. I will double check this is the case, but otherwise I would not know to cite them other than including their source webpage.
>
> I will check and come back to this

Great! The "exit" checklist is above. I'll come back in a few days to follow up.
