+++ paper/paper.md
@@ -10,12 +10,8 @@ tags:
 authors:
   - name: Nicolas T. Kurtovic
     orcid: 0000-0002-2358-4796
-    affiliation: "1, 2" #
-affiliations:
- - name: Max-Planck-Institut f\"{u}r Astronomie, K\"{o}nigstuhl 17, 69117, Heidelberg, Germany.
-   index: 1
- - name: Max-Planck-Institut f\"{u}r Extraterrestrische Physik, Gie{\ss}enbachstra{\ss}e 1, 85748 Garching, Germany.
-   index: 2
+    affiliation: Max-Planck-Institut f\"{u}r Astronomie, K\"{o}nigstuhl 17, 69117, Heidelberg, Germany.
+    affiliation: Max-Planck-Institut f\"{u}r Extraterrestrische Physik, Gie{\ss}enbachstra{\ss}e 1, 85748 Garching, Germany.
 date: 26 March 2024
 bibliography: paper.bib
 
