### 2024-03-26 #4942 @nicokurtovic
@ivastar I fixed the problem with references in line 22. I gave another look at the fakeobs reference, but I could not find a DOI or paper to cite, other than the github repository. I also tried to include multiple affiliations to my name, but I could not do it. I left the affiliation where I did most of the work.

The archive version was uploaded to https://zenodo.org/records/10879411 , with the DOI 10.5281/zenodo.10879411 .
