### 2024-05-08 #4942 @editorialbot
:wave: @openjournals/aass-eics, this paper is ready to be accepted and published.
